# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Example of clean architecture in Go.",
        "title": "Go E-Presciprion Clean API",
        "contact": {},
        "version": "1.0"
    },
    "host": "http://localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/transactions": {
            "get": {
                "description": "Retrieve all transaction records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Get all transactions",
                "operationId": "get-all-transaction",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessTransactionList"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new transaction with medicine details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Create transaction with medicine details",
                "operationId": "create-transaction-with-med-detail",
                "parameters": [
                    {
                        "description": "Transaction input",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateTransaction"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessTransaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/transactions/midtrans-callbacks": {
            "post": {
                "description": "Receives and processes callback notifications from Midtrans",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Handle Midtrans callback notification",
                "operationId": "handle-midtrans-callback",
                "parameters": [
                    {
                        "description": "Midtrans Notification Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessString"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/transactions/patients/{patient_id}": {
            "get": {
                "description": "Retrieve all transactions for a specific patient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Get all transactions by patient ID",
                "operationId": "get-all-transactions-by-patient-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "patient_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessTransactionList"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "description": "Retrieve a single transaction by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Get transaction by ID",
                "operationId": "get-transaction-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessTransaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a transaction by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Delete transaction",
                "operationId": "delete-transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessString"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update a transaction by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction"
                ],
                "summary": "Update transaction",
                "operationId": "update-transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Transaction input",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessTransaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entity.MedicineDetail": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "For pain relief"
                },
                "id": {
                    "type": "string",
                    "example": "1789dfb3-e060-433a-9db4-cb93b42768d8"
                },
                "medicine_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "quantity": {
                    "type": "integer",
                    "example": 2
                },
                "signa_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "transaction_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
...
        "entity.Transaction": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "medicine_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.MedicineDetail"
                    }
                },
                "medicine_type": {
                    "type": "string",
                    "example": "compound"
                },
                "patient_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "payment_redirect_url": {
                    "type": "string",
                    "example": "https://example.com/redirect"
                },
                "payment_token": {
                    "type": "string",
                    "example": "https://example.com/token"
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                },
                "total_medicines": {
                    "type": "integer",
                    "example": 3
                },
                "total_price": {
                    "type": "integer",
                    "example": 15000
                }
            }
        },
        "request.CreateTransaction": {
            "type": "object",
            "required": [
                "descriptions",
                "medicine_type",
                "medicines",
                "patient_id",
                "quantities",
                "signas"
            ],
            "properties": {
                "descriptions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "For pain relief"
                    ]
                },
                "medicine_type": {
                    "type": "string",
                    "enum": [
                        "compound",
                        "non_compound"
                    ],
                    "example": "compound"
                },
                "medicines": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "123e4567-e89b-12d3-a456-426614174000"
                    ]
                },
                "patient_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "quantities": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        2
                    ]
                },
                "signas": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "123e4567-e89b-12d3-a456-426614174000"
                    ]
                }
            }
        },
...
        "request.UpdateTransaction": {
            "type": "object",
            "required": [
                "medicine_type",
                "total_medicines",
                "total_price"
            ],
            "properties": {
                "medicine_type": {
                    "type": "string",
                    "enum": [
                        "compound",
                        "non-compound"
                    ],
                    "example": "compound"
                },
                "total_medicines": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 3
                },
                "total_price": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 15000
                }
            }
        },
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
...
        "response.SuccessString": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "deleted"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.SuccessTransaction": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/entity.Transaction"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.SuccessTransactionList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Transaction"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        }
    }
}